+++ archetypes/schemaextender/tests/testAccessor.py
@@ -1,10 +1,17 @@
-from zope.interface import Interface, implementer, classImplements
-from zope.component import adapts, provideAdapter
-from Products.Archetypes.atapi import ComputedField, StringField, StringWidget
-from Products.ATContentTypes.content.document import ATDocument
-from archetypes.schemaextender.interfaces import ISchemaExtender
+# -*- coding: utf-8 -*-
 from archetypes.schemaextender.field import ExtensionField
-from archetypes.schemaextender.tests.base import ASTestCase as TestCase
+from archetypes.schemaextender.interfaces import ISchemaExtender
+from archetypes.schemaextender.tests.base import ASTestCase
+from plone.testing import zca
+from Products.Archetypes.atapi import ComputedField
+from Products.Archetypes.atapi import StringField
+from Products.Archetypes.atapi import StringWidget
+from Products.ATContentTypes.content.document import ATDocument
+from zope.component import adapter
+from zope.component import provideAdapter
+from zope.interface import classImplements
+from zope.interface import implementer
+from zope.interface import Interface
 
 
 class IFoo(Interface):
@@ -20,19 +27,23 @@ class ExtendedComputedField(ExtensionField, ComputedField):
 
 
 @implementer(ISchemaExtender)
+@adapter(IFoo)
 class Extender(object):
-    adapts(IFoo)
 
     fields = [
-        FooField('foo',
-            widget = StringWidget(label='foo', description='foo!')),
-        FooField('bar',
-            index_method = 'title_or_id',
-            widget = StringWidget(label='bar', description='bar!')),
-        FooField('hmm',
-            index_method = lambda: 'hmm',
-            widget = StringWidget(label='hmm', description='hmm!')),
-        ExtendedComputedField('ho', expression = '"I compute ho"'), ]
+        FooField('foo', widget=StringWidget(label='foo', description='foo!')),
+        FooField(
+            'bar',
+            index_method='title_or_id',
+            widget=StringWidget(label='bar', description='bar!'),
+        ),
+        FooField(
+            'hmm',
+            index_method=lambda: 'hmm',
+            widget=StringWidget(label='hmm', description='hmm!'),
+        ),
+        ExtendedComputedField('ho', expression='"I compute ho"'),
+    ]
 
     def __init__(self, context):
         self.context = context
@@ -41,7 +52,7 @@ class Extender(object):
         return self.fields
 
 
-class AccessorTests(TestCase):
+class AccessorTests(ASTestCase):
 
     def afterSetUp(self):
         classImplements(ATDocument, IFoo)